third_party/blink/web_tests/fast/forms/select/setting-to-invalid-value-expected.txt

A
B
 
A
B
C
D

https://bugs.webkit.org/show_bug.cgi?id=67233 - Setting the value of a select to an invalid value should unset the selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

-- Menu list select:
Setting the value to an invalid value:
PASS sel.selectedIndex is -1
Setting the value to null:
PASS sel.selectedIndex is -1
Setting the value to undefined:
PASS sel.selectedIndex is -1
-- List box select:
Setting the value to an invalid value:
PASS sel.selectedIndex is -1
Setting the value to null:
PASS sel.selectedIndex is -1
Setting the value to undefined:
PASS sel.selectedIndex is -1
PASS successfullyParsed is true

TEST COMPLETE
